requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.7
django-allauth==0.34.0
django-filter==1.1.0
django-oauth-toolkit==1.0.0
django-rest-auth==0.9.2
djangorestframework==3.7.1
idna==2.6
oauthlib==2.0.6
PyJWT==1.5.3
PyMySQL==0.7.11
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
urllib3==1.22
